+++ test/instantiate.jl
@@ -25,8 +25,8 @@ MOI.is_empty(::DummyOptimizer) = true
             MOIU.MockOptimizer{MOIU.UniversalFallback{MOIU.Model{T}}},
         }
         @test MOI.get(optimizer, MOI.Silent())
-        @test MOI.get(optimizer, MOI.RawParameter("a")) == 1
-        @test MOI.get(optimizer, MOI.RawParameter("b")) == 2
+        @test MOI.get(optimizer, MOI.RawOptimizerAttribute("a")) == 1
+        @test MOI.get(optimizer, MOI.RawOptimizerAttribute("b")) == 2
     end
 
     optimizer_constructor = MOI.OptimizerWithAttributes(DummyOptimizer, [])